Pull caching logic into transforms
